Move extensions/helpers to membership-common
