+++ components/membership/membership-group-read-impl/src/main/kotlin/net/corda/membership/impl/read/reader/MembershipGroupReaderImpl.kt
@@ -1,7 +1,7 @@
 package net.corda.membership.impl.read.reader
 
-import net.corda.membership.lib.impl.MemberInfoExtension.Companion.ledgerKeyHashes
-import net.corda.membership.lib.impl.MemberInfoExtension.Companion.sessionKeyHash
+import net.corda.membership.lib.MemberInfoExtension.Companion.ledgerKeyHashes
+import net.corda.membership.lib.MemberInfoExtension.Companion.sessionKeyHash
 import net.corda.membership.lib.CPIWhiteList
 import net.corda.membership.impl.read.cache.MembershipGroupReadCache
 import net.corda.membership.read.MembershipGroupReader
