+++ components/virtual-node/virtual-node-write-service-impl/build.gradle
@@ -23,7 +23,6 @@ dependencies {
     implementation project(':libs:db:db-orm')
     implementation project(':libs:lifecycle:lifecycle')
     implementation project(":libs:membership:membership-common")
-    implementation project(":libs:membership:membership-impl")
     implementation project(":libs:messaging:messaging")
     implementation project(":libs:utilities")
     implementation project(":libs:virtual-node:cpi-datamodel")
