+++ libs/membership/ledger-identity/src/test/kotlin/net/corda/membership/ledger/identity/converter/PartyConverterTest.kt
@@ -4,10 +4,10 @@ import net.corda.layeredpropertymap.create
 import net.corda.layeredpropertymap.testkit.LayeredPropertyMapMocks
 import net.corda.membership.ledger.identity.PartyImpl
 import net.corda.membership.lib.impl.MemberContextImpl
-import net.corda.membership.lib.impl.MemberInfoExtension.Companion.NOTARY_SERVICE_SESSION_KEY
-import net.corda.membership.lib.impl.MemberInfoExtension.Companion.NOTARY_SERVICE_PARTY_NAME
-import net.corda.membership.lib.impl.MemberInfoExtension.Companion.PARTY_NAME
-import net.corda.membership.lib.impl.MemberInfoExtension.Companion.PARTY_SESSION_KEY
+import net.corda.membership.lib.MemberInfoExtension.Companion.NOTARY_SERVICE_SESSION_KEY
+import net.corda.membership.lib.MemberInfoExtension.Companion.NOTARY_SERVICE_PARTY_NAME
+import net.corda.membership.lib.MemberInfoExtension.Companion.PARTY_NAME
+import net.corda.membership.lib.MemberInfoExtension.Companion.PARTY_SESSION_KEY
 import net.corda.v5.base.types.MemberX500Name
 import net.corda.v5.cipher.suite.CipherSchemeMetadata
 import net.corda.v5.ledger.obsolete.identity.Party
