+++ libs/membership/membership-common/src/main/kotlin/net/corda/membership/lib/MemberInfoExtension.kt
@@ -1,4 +1,4 @@
-package net.corda.membership.lib.impl
+package net.corda.membership.lib
 
 import net.corda.v5.base.util.NetworkHostAndPort
 import net.corda.v5.base.util.parse
